Crappy solution

+++ Day18/Position.cs
@@ -1,3 +1,3 @@
 namespace Day18;
 
-public record Position(int X, int Y);
+public record Position(double X, double Y);
